task 4.2.2 + файлы констант
