--- lib/constants.dart
@@ -1,11 +1,0 @@
-import 'package:flutter/material.dart';
-
-const kMyToolbarHeight = 112.0;
-const kAppBarElevation = 0.0;
-const kTitlePaddingTop = 40.0;
-const title1 = "С";
-const title2 = "писок\n";
-const title3 = "и";
-const title4 = "нтересных мест";
-const kTextColor = Color(0xFF3B3E5B);
-const kAppBarFontSize = 32.0;
