+++ lib/ui/res/colors.dart
@@ -1,0 +1,3 @@
+import 'dart:ui';
+
+const kTextColor = Color(0xFF3B3E5B);
