+++ lib/ui/res/constants.dart
@@ -1,0 +1,4 @@
+const kMyToolbarHeight = 112.0;
+const kAppBarElevation = 0.0;
+const kTitlePaddingTop = 40.0;
+const kAppBarFontSize = 32.0;
