+++ lib/ui/res/strings.dart
@@ -1,0 +1,4 @@
+const title1 = "С";
+const title2 = "писок\n";
+const title3 = "и";
+const title4 = "нтересных мест";
