+++ lib/ui/res/text_styles.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+import 'colors.dart';
+import 'constants.dart';
+
+const kMainAppBarTextStyle =
+    TextStyle(color: kTextColor, fontSize: kAppBarFontSize);
+
+const kGreenAppBarTextStyle = TextStyle(color: Colors.green, fontSize: kAppBarFontSize);
+
+const kYellowAppBarTextStyle = TextStyle(color: Colors.yellow, fontSize: kAppBarFontSize);
